requirements.txt
django==5.2.6
djangorestframework
djangorestframework-simplejwt
django-redis
redis
drf-spectacular
django-filter
celery[redis]==5.4.0
flower==2.0.1
psycopg[binary]==3.1.18
requests>=2.31
Pillow>=10.0.0
minio>=7.1.0
